Update docker entrypoint (#25)
* create /App/Frierun in entrypoint if it doesn't exist
* create empty state.json if no docker socket found
* update README.md for using the `latest` tag
* set max heap size to 200Mb. It should be enough for Frierun and will fix running it on Termux

+++ entrypoint.sh
@@ -2,9 +2,17 @@
 
 set -e
 
+if [ ! -d /App/Frierun ]; then
+  mkdir -p /App/Frierun
+fi
+
 if [ ! -f /App/Frierun/state.json ]; then
   echo "No state.json file found, creating one"
-  cp /App/state.json.default /App/Frierun/state.json
+  if [ -f /var/run/docker.sock ]; then 
+    cp /App/state.json.default /App/Frierun/state.json
+  else
+    touch /App/Frierun/state.json
+  fi
 fi
 
 exec "$@"
